+++ stdlib/public/core/CollectionOfOne.swift
@@ -136,14 +136,17 @@ extension CollectionOfOne: RandomAccessCollection, MutableCollection {
   }
 }
 
+@available(swift, introduced: 4.1) // FIXME(conformance-availability)
 extension CollectionOfOne : Equatable where Element : Equatable {
   /// Returns a Boolean value indicating whether two collections are equal.
   @_inlineable // FIXME(sil-serialize-all)
+  @available(swift, introduced: 4.1)
   public static func == (lhs: CollectionOfOne, rhs: CollectionOfOne) -> Bool {
     return lhs._element == rhs._element
   }
 }
 
+@available(swift, introduced: 4.1) // FIXME(conformance-availability)
 extension CollectionOfOne : Hashable where Element : Hashable {
   /// The hash value for the collection.
   ///
@@ -152,6 +155,7 @@ extension CollectionOfOne : Hashable where Element : Hashable {
   /// Hash values are not guaranteed to be equal across different executions of
   /// your program. Do not save hash values to use during a future execution.
   @_inlineable // FIXME(sil-serialize-all)
+  @available(swift, introduced: 4.1)
   public var hashValue: Int {
     return _element.hashValue
   }
