+++ stdlib/public/core/EmptyCollection.swift
@@ -173,8 +173,10 @@ extension EmptyCollection : Equatable {
   }
 }
 
+@available(swift, introduced: 4.1) // FIXME(conformance-availability)
 extension EmptyCollection : Hashable {
   @_inlineable // FIXME(sil-serialize-all)
+  @available(swift, introduced: 4.1)
   public var hashValue: Int {
     return 0
   }
