+++ stdlib/public/core/Mirror.swift
@@ -905,8 +905,10 @@ extension DictionaryLiteral : RandomAccessCollection {
   }
 }
 
+@available(swift, introduced: 4.1) // FIXME(conformance-availability)
 extension DictionaryLiteral: Equatable where Key: Equatable, Value: Equatable {
   @_inlineable // FIXME(sil-serialize-all)
+  @available(swift, introduced: 4.1)
   public static func == (
     lhs: DictionaryLiteral<Key, Value>, rhs: DictionaryLiteral<Key, Value>
   ) -> Bool {
@@ -917,6 +919,7 @@ extension DictionaryLiteral: Equatable where Key: Equatable, Value: Equatable {
   }
 }
 
+@available(swift, introduced: 4.1) // FIXME(conformance-availability)
 extension DictionaryLiteral: Hashable where Key: Hashable, Value: Hashable {
   /// The hash value for the collection.
   ///
@@ -926,6 +929,7 @@ extension DictionaryLiteral: Hashable where Key: Hashable, Value: Hashable {
   /// Hash values are not guaranteed to be equal across different executions of
   /// your program. Do not save hash values to use during a future execution.
   @_inlineable // FIXME(sil-serialize-all)
+  @available(swift, introduced: 4.1)
   public var hashValue: Int {
     // FIXME(ABI)#177: <rdar://problem/18915294> Issue applies to DictionaryLiteral too
     var result: Int = 0
